Using CTE, JOIN, Sub-Queries, Aggregate and WHERE functions
While finding the id, name, and total count of submissions is basic - it becomes an issue if you want to take the maximum number of submissions and weed out "ties" for submissions under that count.

This is because the WHERE clause does not allow the use of aggregate functions.  To bypass this, we use a sub-query to locate all tied submissions.  However, we then must create a sub-query within that in order to define the maximum count of submissions.  Once this is done, we can now easily pull the content we need to exclude.

A CTE was used as the same table was referenced multiple times and would get the most mileage in use and efficiency.