Example of advanced usage of CTE's and Window Functions.
Utilized CTE's and Window Functions (RANK, LAG, and FIRST_VALUE) to calculate the total income for the weekday and compare the percentile difference to the day prior to it.  Additionally, we also display a percentile difference between that day and the highest earning day as well - and rank them accordingly.